requirements.txt

httpx
pytest
allure-python-commons
pydantic-settings
pytest-asyncio
pydantic
faker